refactor: add opt out messaging to drawer template, extract opt out logic into separate component, ensure welcome message styles stay intact with extra buttons
