+++ js-setup/js/component-initializer.js
@@ -3,6 +3,7 @@ import ads from '../../ads';
 import tracking from '../../tracking';
 import date from '../../date';
 import header from '../../header';
+import optInOut from '../../opt-in-out';
 import promoMessages from '../../promo-messages';
 import cookieMessage from '../../cookie-message';
 import welcomeMessage from '../../welcome-message';
@@ -79,6 +80,10 @@ export class ComponentInitializer {
 				serviceWorker.register(flags);
 			} else {
 				serviceWorker.unregister();
+			}
+
+			if (flags.get('optInOut')) {
+				optInOut.init();
 			}
 
 			if (config.features.myft && !this.initializedFeatures.myftclient) {
