+++ normalize/main.scss
@@ -222,7 +222,7 @@
 	*/
 
 	[hidden] {
-		display: none;
+		display: none !important;
 	}
 
 	/* Next specific stuff
