+++ welcome-message/main.scss
@@ -18,10 +18,6 @@
 		@media print {
 			display: none;
 		}
-
-		&.is-hidden {
-			display: none;
-		}
 	}
 
 	.n-welcome-message--static {
@@ -52,7 +48,7 @@
 	}
 
 	.n-welcome-message__column {
-		@include oGridRespondTo('M') {
+		@include oGridRespondTo('L') {
 			display: table-cell;
 			vertical-align: middle;
 		}
@@ -63,10 +59,8 @@
 	.n-welcome-message__column--actions {
 		white-space: nowrap;
 
-		@include oGridRespondTo('M') {
-			width: 35%;
-			// in combination with nowrap it will force elements to flow from right
-			direction: rtl;
+		@include oGridRespondTo('L') {
+			text-align: right;
 		}
 	}
 
@@ -83,22 +77,18 @@
 			margin: 0 0 12px;
 		}
 
-		@include oGridRespondTo('M') {
+		@include oGridRespondTo('L') {
 			margin-bottom: 0;
 		}
 	}
 
 	.n-welcome-message__button {
 		@include nButtons($theme:'standout');
-		margin-left: 12px;
+		margin-left: 6px;
 		white-space: nowrap;
 
 		&:first-child {
 			margin-left: 0;
-		}
-
-		&.is-hidden {
-			display: none;
 		}
 	}
