+++ welcome-message/template.html
@@ -4,7 +4,7 @@
 	</div>
 </div>
 
-<div class="n-welcome-message n-welcome-message--fixed is-hidden" data-trackable="welcome">
+<div class="n-welcome-message n-welcome-message--fixed" data-trackable="welcome" hidden>
 	<div class="o-grid-container">
 		{{>n-ui/welcome-message/message}}
 	</div>
